requirements.txt
jax
jaxlib
pytest
tensorboardX
hydra-core
flax
